# src/backend/api/docs/swagger.yaml
basePath: /api
definitions:
...
  model.ChatMessage:
    properties:
      chat_session_id:
        type: number
      citations:
        items:
          $ref: '#/definitions/model.DocumentResponse'
        type: array
      error:
        type: string
      feedback:
        $ref: '#/definitions/model.ChatMessageFeedback'
      id:
        type: number
      latest_child_message:
        type: integer
      message:
        type: string
      message_type:
        type: string
      parent_message:
        type: number
      rephrased_query:
        type: string
      time_sent:
        type: string
      token_count:
        type: integer
    type: object
  model.ChatMessageFeedback:
    properties:
      chat_message_id:
        type: number
      feedback:
        type: string
      id:
        type: number
      up_votes:
        type: boolean
      user_id:
        type: string
    type: object
  model.ChatSession:
    properties:
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      description:
        type: string
      id:
        type: number
      messages:
        items:
          $ref: '#/definitions/model.ChatMessage'
        type: array
      one_shot:
        type: boolean
      persona:
        $ref: '#/definitions/model.Persona'
      persona_id:
        type: number
      user_id:
        type: string
    type: object
...
  model.DocumentResponse:
    properties:
      content:
        type: string
      document_id:
        type: string
      id:
        type: number
      link:
        type: string
      message_id:
        type: number
      updated_date:
        type: string
    type: object
...
  model.LLM:
    properties:
      api_key:
        type: string
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      endpoint:
        type: string
      id:
        type: number
      last_update:
        $ref: '#/definitions/pg.NullTime'
      model_id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      url:
        type: string
    type: object
  model.Persona:
    properties:
      chat_sessions:
        items:
          $ref: '#/definitions/model.ChatSession'
        type: array
      creation_date:
        type: string
      default_persona:
        type: boolean
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      description:
        type: string
      display_priority:
        type: integer
      id:
        type: number
      is_visible:
        type: boolean
      last_update:
        $ref: '#/definitions/pg.NullTime'
      llm:
        $ref: '#/definitions/model.LLM'
      llm_id:
        type: number
      name:
        type: string
      prompt:
        $ref: '#/definitions/model.Prompt'
      starter_messages:
        items:
          type: integer
        type: array
      tenant_id:
        type: string
    type: object
  model.Prompt:
    properties:
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      description:
        type: string
      id:
        type: number
      last_update:
        $ref: '#/definitions/pg.NullTime'
      name:
        type: string
      persona_id:
        type: number
      system_prompt:
        type: string
      task_prompt:
        type: string
      user_id:
        type: string
    type: object
...
  parameters.CreateChatMessageRequest:
    properties:
      chat_session_id:
        type: number
      message:
        type: string
      no_ai_answer:
        type: boolean
      parent_message_id:
        type: number
      prompt_id:
        type: number
      query_override:
        description: RetrievalOptions RetrievalDetails  `json:"retrieval_options,omitempty"`
        type: string
      search_doc_ids:
        items:
          type: number
        type: array
    type: object
  parameters.CreateChatSession:
    properties:
      description:
        type: string
      one_shot:
        type: boolean
      persona_id:
        type: number
    type: object
...
  parameters.MessageFeedbackParam:
    properties:
      id:
        type: number
      vote:
        type: string
    type: object
...
  pg.NullTime:
    properties:
      time.Time:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: This is Cognix Golang API Documentation
  title: Cognix API
  version: "1.0"
paths:
  /auth/google/login:
    get:
      description: login using google auth
      operationId: auth_login
      parameters:
      - description: redirect base url
        in: query
        name: redirect_url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: login using google auth
      tags:
      - Auth
  /chats/create-chat-session:
    post:
      description: creates new chat session
      operationId: chat_create_session
      parameters:
      - description: create session parameters
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/parameters.CreateChatSession'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChatSession'
      security:
      - ApiKeyAuth: []
      summary: creates new chat session
      tags:
      - Chat
  /chats/get-chat-session/{id}:
    get:
      description: return chat session with messages by given id
      operationId: chat_get_by_id
      parameters:
      - description: session id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChatSession'
      security:
      - ApiKeyAuth: []
      summary: return chat session with messages by given id
      tags:
      - Chat
  /chats/get-user-chat-sessions:
    get:
      description: return list of chat sessions for current user
      operationId: chat_get_sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ChatSession'
            type: array
      security:
      - ApiKeyAuth: []
      summary: return list of chat sessions for current user
      tags:
      - Chat
  /chats/message/feedback:
    post:
      description: add  feedback to message
      operationId: chat_message_feedback
      parameters:
      - description: send message parameters
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/parameters.MessageFeedbackParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChatMessageFeedback'
      security:
      - ApiKeyAuth: []
      summary: add  feedback to message
      tags:
      - Chat
  /chats/send-message:
    post:
      description: send message and wait stream response
      operationId: chat_send_message
      parameters:
      - description: send message parameters
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/parameters.CreateChatMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChatMessage'
      security:
      - ApiKeyAuth: []
      summary: send message and wait stream response
      tags:
      - Chat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"